# openapi/polyglot.yaml
openapi: 3.1.0
info:
  title: polyglot Data API
  version: "0.2.0"
  description: |
    polyglot is val-analyzer's generic Data API host: it owns a small PocketBase-embedded
    onboarding/catalog database (which datasources exist, their tables/columns/background actions,
    and any hand-curated descriptions or AI query guidance) and exposes a small, engine-agnostic REST
    contract over every onboarded datasource - submit ANSI-compliant SQL to read cached data,
    introspect the live catalog, trigger a datasource's own background data-refresh actions, and
    curate descriptions/guidance.

    polyglot has no AI/reasoning logic of its own - it is a pure data API. A separate MCP server
    wraps these endpoints as tools for an AI model to call when answering a user's question.

    polyglot also has no built-in domain knowledge and no data of its own beyond its onboarding
    bookkeeping: every table a caller can query comes from a `dataprovider.Provider` connection
    (see `internal/dataprovider`) onboarded at runtime via `POST /datasources` - either a local
    SQLite file (`sqlite`), or another service speaking this same small query contract over the
    network (`http_sql`). Valorant's own data lives entirely in a separate standalone service,
    `cmd/valorantapi` (its own PocketBase, its own `/query`+`/schema`+`/functions`+`/warm`), onboarded
    into polyglot as an ordinary `http_sql` datasource named `valorant` - not a special case.
...
servers:
...
security:
  - bearerAuth: []
paths:
  /query:
    get:
      operationId: query
      summary: Run a read-only ANSI SQL query
      description: |
        Executes a single `SELECT` or `WITH` statement and returns matching rows as JSON objects
        keyed by column name.

        IMPORTANT - almost every call needs `datasource` set: every table returned by
        GET /metadata (e.g. `players`, `matches`, ...) belongs to a specific onboarded datasource,
        tagged in that response as `tables[].datasource`. To query any of those tables, you MUST
        pass that same value as `datasource` on this call. Omitting `datasource` does NOT search
        across all onboarded datasources or fall back to one automatically - it queries a
        completely different, separate database: polyglot's own onboarding/catalog bookkeeping
        (the `datasources`/`tables`/`columns` administrative tables, not data tables - and even
        those are blocked from being read here). If you omit `datasource` and query e.g. `players`,
        you will get a "no such table" error, not the Valorant (or any other datasource's) data -
        that error means you forgot `datasource`, not that the table doesn't exist. Check
        GET /metadata first if you're unsure which `datasource` value a table belongs to.

        Each datasource is queried through its own separate connection, so a `datasource`-scoped
        query never joins across datasources or against polyglot's own bookkeeping tables.

        The query runs against a read-only connection as the primary safety boundary; the
        SELECT/WITH prefix check is defense in depth on top of that, not the main guarantee.
        Include your own `LIMIT`/`OFFSET` in the SQL text - polyglot does not layer separate
        pagination parameters on top of the query, it only enforces a hard server-side cap on rows
        returned (see `truncated` in the response).
      parameters:
        - name: sql
          in: query
          required: true
          schema:
            type: string
          description: An ANSI-compliant SQL `SELECT` or `WITH` statement, to run against `datasource` if given.
          example: "SELECT name, tag, region FROM players WHERE tag = 'NA1' LIMIT 10"
        - name: datasource
          in: query
          required: false
          schema:
            type: string
          description: |
            Name of an onboarded datasource (see GET /metadata's tables[].datasource or
            GET /datasources' active[].name) to query. Required to reach any actual data table
            (players, matches, or any other table GET /metadata lists) - omitting this does not
            mean "search everywhere," it means "query polyglot's own internal bookkeeping instead,"
            which has none of that data.
          example: valorant
      responses:
        "200":
          description: Query executed successfully.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/QueryResponse"
              example:
                rows:
                  - puuid: "abc123"
                    name: "OrBest"
                    tag: "NA1"
                row_count: 1
                truncated: false
        "400":
          description: The query is missing, empty, not a SELECT/WITH statement, references a reserved bookkeeping table, or names an unknown datasource.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "500":
          description: The query failed to execute (syntax error, unknown table/column, timeout).
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
  /metadata:
    get:
      operationId: getMetadata
      summary: Describe every onboarded datasource's tables/columns/functions and curated guidance
      description: |
        Returns polyglot's persisted catalog snapshot - every onboarded datasource's
        connection-level description/query_guidance, every table's columns/types/hand-curated
        description/query_guidance, and every background action that datasource exposes (if any -
        most don't). This is built from a snapshot kept current by an async reconciliation pass (see
        POST /datasources and POST /datasources/reconcile), not a live call into every datasource, so
        this endpoint's latency never depends on any one datasource's health or speed - it's the
        discovery mechanism an MCP client or AI model uses to know what it can query (and what it can
        trigger via POST /warm), without reading polyglot's source or waiting on a slow/unreachable
        connection.

        `query_guidance` is distinct from `description`: description says what a table/connection
        means, query_guidance says how to query it well (e.g. "this connection is a columnar OLAP
        warehouse - always filter on the partition column" at the datasource level, or "partitioned
        by event_date - always filter on it" at the table level). Both start empty until curated via
        POST /datasources/annotate, POST /tables/annotate, or POST /columns/annotate.

        `functions[]` lists whatever named background actions the datasource itself exposes (e.g. the
        Valorant datasource's `sync_matches`) - each entry's `name`/`datasource`/`args` are exactly
        what POST /warm expects. `description` here is not hand-curated (it always mirrors what the
        datasource itself reports and refreshes every reconcile); only `query_guidance` is ever
        curated, via POST /functions/annotate.
      responses:
        "200":
          description: Current catalog snapshot.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MetadataResponse"
              example:
                datasources:
                  - name: valorant
                    description: "Valorant esports match/player data."
                    query_guidance: ""
                tables:
                  - id: "rec_abc123"
                    name: players
                    description: "Cached Riot account identities."
                    datasource: valorant
                    query_guidance: ""
                    columns:
                      - id: "rec_def456"
                        name: puuid
                        type: TEXT
                        description: "Riot's globally unique player id."
                      - id: "rec_ghi789"
                        name: name
                        type: TEXT
                        description: "Riot ID name, before the #."
                functions:
                  - id: "rec_jkl012"
                    name: sync_matches
                    description: "Fetch and cache a player's matches from the upstream Valorant API..."
                    datasource: valorant
                    query_guidance: ""
                    args:
                      - name: player_tag
                        type: string
                        description: "The player's Riot ID as name#tag, e.g. \"Orbest#NA1\"."
                        required: true
        "401":
          $ref: "#/components/responses/Unauthorized"
  /datasources:
    get:
      operationId: listDatasources
      summary: List available provider types and active datasources
      description: |
        Describes every compiled-in provider type (its config schema, for onboarding) and every
        currently active datasource's name and provider type. This is the discovery mechanism for
        what can be onboarded and what already has been, without reading polyglot's source - see
        GET /metadata for the actual tables/columns each active datasource exposes.
      responses:
        "200":
          description: Current provider types and active datasources.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/DatasourcesResponse"
              example:
                available_types:
                  - type: sqlite
                    config:
                      - name: path
                        type: string
                        required: true
                        secret: false
                        description: "Absolute filesystem path to the SQLite database file, readable by the polyglot process."
                  - type: http_sql
                    config:
                      - name: base_url
                        type: string
                        required: true
                        secret: false
                        description: "Base URL of a service exposing GET /query and GET /schema."
                      - name: auth_token
                        type: string
                        required: true
                        secret: true
                        description: "Bearer token for the remote service."
                active:
                  - name: valorant
                    type: http_sql
        "401":
          $ref: "#/components/responses/Unauthorized"
    post:
      operationId: onboardDatasource
      summary: Onboard (or reconfigure) a datasource under a caller-chosen name
      description: |
        Validates config against the named provider type's config schema, connects (a real dial/
        open, not just a config check - a bad connection fails onboarding immediately, not the
        first query), persists the registration (any config field marked `secret` is written to
        OpenBao and replaced with a vault path reference before being stored - see the root
        CLAUDE.md), activates it under `name`, and kicks off an async catalog-reconciliation pass
        (see `reconcile_job_id` in the response, pollable via GET /jobs?id=). Re-onboarding an
        already-active name is an idempotent upsert: its stored config is replaced and a fresh
        instance replaces the old one. `name` is a caller-chosen identifier, independent of
        `type` - multiple datasources can share one provider type (e.g. two separate onboarded
        SQLite files).
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/OnboardRequest"
            example:
              name: valorant
              type: http_sql
              config:
                base_url: "http://valorantapi:8093"
                auth_token: "..."
      responses:
        "200":
          description: Datasource onboarded successfully.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/OnboardResponse"
              example:
                name: valorant
                type: http_sql
                config:
                  base_url: "http://valorantapi:8093"
                  auth_token:
                    $vault: "datasources/valorant/auth_token"
                reconcile_job_id: "aB3dEfGhIjKlMnO"
        "400":
          description: Unknown provider type, invalid/missing required config field, or a reserved name (datasources, tables, columns).
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "401":
          $ref: "#/components/responses/Unauthorized"
  /datasources/reconcile:
    post:
      operationId: reconcileDatasource
      summary: Re-run catalog reconciliation for an already-onboarded datasource
      description: |
        Re-introspects the named datasource's live schema and upserts polyglot's persisted
        tables/columns snapshot: new tables/columns are added, ones no longer present are removed,
        and a column's type is refreshed - but existing hand-curated description/query_guidance
        text is never touched. Runs asynchronously (this is also triggered automatically after
        every successful POST /datasources onboard - this endpoint is for re-running it later,
        e.g. after the underlying schema changed).
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ReconcileRequest"
            example:
              name: valorant
      responses:
        "202":
          description: Reconciliation job accepted and running in the background.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Job"
              example:
                id: "aB3dEfGhIjKlMnO"
                datasource: valorant
                function: reconcile_catalog
                status: running
                created_at: "2026-07-12T18:04:05Z"
                updated_at: "2026-07-12T18:04:05Z"
        "400":
          description: Missing name, or unknown datasource.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "401":
          $ref: "#/components/responses/Unauthorized"
  /datasources/annotate:
    post:
      operationId: annotateDatasource
      summary: Curate a datasource's connection-level description/query_guidance
      description: |
        Patches only the fields present in the request body - omit a field to leave it unchanged,
        or send an empty string to explicitly clear it. `query_guidance` here is connection-level
        AI guidance distinct from any one table's own (e.g. "this connection is a columnar OLAP
        warehouse - queries without a partition filter will be slow").
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AnnotateDatasourceRequest"
            example:
              name: valorant
              description: "Valorant esports match/player data, synced from the HenrikDev API."
              query_guidance: "matches.raw_json is large - prefer the typed columns/tables for aggregate queries."
      responses:
        "200":
          description: Annotation saved.
          content:
            application/json:
              schema:
                type: object
                required: [name]
                properties:
                  name:
                    type: string
        "400":
          description: Missing name.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "404":
          description: Unknown datasource name.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
  /tables/annotate:
    post:
      operationId: annotateTable
      summary: Curate one table's description/query_guidance
      description: |
        Patches only the fields present in the request body. `id` is a table's id from
        GET /metadata's `tables[].id`.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AnnotateTableRequest"
            example:
              id: "rec_abc123"
              description: "One row per cached Valorant match."
              query_guidance: ""
      responses:
        "200":
          description: Annotation saved.
          content:
            application/json:
              schema:
                type: object
                required: [id]
                properties:
                  id:
                    type: string
        "400":
          description: Missing id.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "404":
          description: Unknown table id.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
  /columns/annotate:
    post:
      operationId: annotateColumn
      summary: Curate one column's description
      description: |
        Patches the description field if present in the request body. No query_guidance field
        here (unlike tables/annotate) - table-level guidance can name specific columns when
        relevant; a third tier would be over-engineering. `id` is a column's id from
        GET /metadata's `tables[].columns[].id`.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AnnotateColumnRequest"
            example:
              id: "rec_def456"
              description: "Stable Riot player identifier; the true identity key (name/tag can change)."
      responses:
        "200":
          description: Annotation saved.
          content:
            application/json:
              schema:
                type: object
                required: [id]
                properties:
                  id:
                    type: string
        "400":
          description: Missing id.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "404":
          description: Unknown column id.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
  /warm:
    post:
      operationId: warm
      summary: Trigger a named background data-refresh action on an onboarded datasource
      description: |
        Proxies a named-function trigger through to `datasource`'s own background-action capability
        (not every datasource has one - most don't). This is how a fresh sync gets triggered - e.g.
        pulling a Valorant player's latest matches from the upstream HenrikDev API - since plain
        `query` only ever reads whatever is already cached.

        Find valid `datasource`/`function`/`args` values via GET /metadata's `functions[]` list:
        each entry's `datasource` and `name` are exactly what to pass here, and `args[]` documents
        what that function accepts (name/type/description/required).

        IMPORTANT - only call this when a human has explicitly asked you to refresh, update, or sync
        data. Never call it just because a `query` came back empty or incomplete: this endpoint is
        asynchronous and returns a `Job` immediately, long before the actual work finishes - it will
        not produce usable data in time to help answer the current question. If the data you need
        isn't cached, say so plainly instead of trying to warm it yourself mid-question. Poll the
        returned job via `GET /jobs?id=&datasource=` if you need to confirm it completed, e.g. because
        the user is waiting on that specifically.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/WarmRequest"
            example:
              datasource: valorant
              function: sync_matches
              args:
                player_tag: "OrBest#NA1"
      responses:
        "202":
          description: Accepted and running in the background.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Job"
              example:
                id: "aB3dEfGhIjKlMnO"
                datasource: valorant
                function: sync_matches
                status: running
                created_at: "2026-07-23T01:23:15Z"
                updated_at: "2026-07-23T01:23:15Z"
        "400":
          description: Missing datasource/function, unknown datasource, the datasource doesn't support background actions, or the datasource itself rejected the call (e.g. unknown function name or a missing required arg).
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "401":
          $ref: "#/components/responses/Unauthorized"
  /functions/annotate:
    post:
      operationId: annotateFunction
      summary: Curate one function's query_guidance
      description: |
        Patches the query_guidance field if present in the request body. No description field here
        (unlike tables/annotate/datasources/annotate) - a function's description already comes from
        the datasource itself and is refreshed automatically every reconcile, so there's nothing to
        hand-curate there; query_guidance is for extra operational notes beyond that (e.g. "never set
        full_history without explicit user confirmation - it can take a very long time"). `id` is a
        function's id from GET /metadata's `functions[].id`.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AnnotateFunctionRequest"
            example:
              id: "rec_jkl012"
              query_guidance: "Only pass full_history=true if the user explicitly asks for a player's entire match history - it can take several minutes."
      responses:
        "200":
          description: Annotation saved.
          content:
            application/json:
              schema:
                type: object
                required: [id]
                properties:
                  id:
                    type: string
        "400":
          description: Missing id.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "404":
          description: Unknown function id.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
  /jobs:
    get:
      operationId: getJob
      summary: Check the status/result of a background job
      description: |
        Returns the current status of a job: `running`, `succeeded` (with `summary`/`data`
        populated), or `failed` (with `error` populated).

        Two different kinds of job share this one polling endpoint, distinguished by whether
        `datasource` is given:
        - Omit `datasource` for a job started by POST /datasources (the automatic post-onboard
          reconcile) or POST /datasources/reconcile - checked against polyglot's own local job
          tracking.
        - Pass `datasource` for a job started by POST /warm - the poll is proxied straight through
          to that datasource's own job tracking (e.g. the standalone Valorant API's own /warm), so
          it must be the same `datasource` value the POST /warm call used.
      parameters:
        - name: id
          in: query
          required: true
          schema:
            type: string
          description: A job id returned by POST /datasources, POST /datasources/reconcile, or POST /warm.
          example: "aB3dEfGhIjKlMnO"
        - name: datasource
          in: query
          required: false
          schema:
            type: string
          description: |
            Required if, and only if, `id` came from POST /warm - name of the datasource that job
            was started against (the same value passed to POST /warm). Omit entirely for a
            POST /datasources / POST /datasources/reconcile job id - passing a datasource for one of
            those is not meaningful and will fail to find the job.
          example: valorant
      responses:
        "200":
          description: The job exists; check `status` for whether it's finished.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Job"
              example:
                id: "aB3dEfGhIjKlMnO"
                datasource: valorant
                function: reconcile_catalog
                status: succeeded
                summary: "reconciled catalog for \"valorant\""
                created_at: "2026-07-12T18:04:05Z"
                updated_at: "2026-07-12T18:04:41Z"
        "400":
          description: Missing id query parameter, or datasource is set but unknown/doesn't support background actions.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "404":
          description: No job with that id is currently tracked (unknown id, or its result was evicted after finishing) - for a datasource-proxied job, this reflects that datasource's own job tracking, not polyglot's.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
  /queries:
    get:
      operationId: listQueries
      summary: List recently run SQL queries
      description: |
        Returns the most recent SQL queries actually run through GET /query (across every caller -
        an MCP client, a direct curl call, anything), most recent first. Read live from Jaeger's own
        trace storage rather than a persisted collection of polyglot's own - so this list is empty if
        tracing isn't configured, and resets whenever Jaeger itself restarts (its storage is
        deliberately ephemeral/in-memory, by this project's own design).

        One row per actual query call, not per conversation - a single multi-turn AI conversation can
        run several queries, and each one gets its own row here.
      responses:
        "200":
          description: The most recent queries, most recent first.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ListQueriesResponse"
              example:
                queries:
                  - id: "5326331314ad6a7fd748441a2e63290e:fcdf65ab6774b02a"
                    sql: "SELECT name, tag FROM players WHERE tag = 'NA1' LIMIT 10"
                    question: "how good is Sova this act?"
                    status: success
                    duration_ms: 12
                    timestamp: "2026-07-12T18:04:05Z"
        "401":
          $ref: "#/components/responses/Unauthorized"
  /queries/detail:
    get:
      operationId: getQueryDetail
      summary: Get the full detail of one previously run query
      description: |
        Returns everything GET /queries' summary row has, plus the query's full response body and
        every span in the same trace (one entry per hop it passed through - e.g. mcpserver, polyglot,
        valorantapi - each with its own timing and tags) for full cross-service context.
      parameters:
        - name: id
          in: query
          required: true
          schema:
            type: string
          description: A query id from GET /queries' `id` field (opaque, "<traceID>:<spanID>" shaped - do not construct one by hand).
          example: "5326331314ad6a7fd748441a2e63290e:fcdf65ab6774b02a"
      responses:
        "200":
          description: The query exists; full detail follows.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/QueryDetail"
        "400":
          description: Missing or malformed id query parameter.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "404":
          description: No query with that id (unknown id, or Jaeger's own storage has already evicted/rotated it out).
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      description: |
        Static shared-secret bearer token, constant-time compared server-side. Not a per-user
        identity.
  responses:
    Unauthorized:
      description: Missing or invalid bearer token.
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
  schemas:
    Error:
      type: object
      description: PocketBase's standard API error envelope.
      required: [status, message, data]
      properties:
        status:
          type: integer
          example: 400
        message:
          type: string
          example: "invalid request body"
        data:
          type: object
          additionalProperties: true
...
    QueryResponse:
      type: object
      required: [rows, row_count, truncated]
      properties:
        rows:
          type: array
          items:
            $ref: "#/components/schemas/Row"
        row_count:
          type: integer
          minimum: 0
          description: Number of rows in `rows`.
        truncated:
          type: boolean
          description: True if the server-side row safety cap cut off further results.
...
    MetadataResponse:
      type: object
      required: [datasources, tables, functions]
      properties:
        datasources:
          type: array
          items:
            $ref: "#/components/schemas/DatasourceGuidance"
        tables:
          type: array
          items:
            $ref: "#/components/schemas/TableDescription"
        functions:
          type: array
          items:
            $ref: "#/components/schemas/FunctionDescription"
...
    DatasourcesResponse:
      type: object
      required: [available_types, active]
      properties:
        available_types:
          type: array
          items:
            $ref: "#/components/schemas/ProviderTypeDescription"
        active:
          type: array
          items:
            $ref: "#/components/schemas/ActiveDatasourceDescription"
    OnboardRequest:
      type: object
      required: [name, type]
      properties:
        name:
          type: string
          description: Caller-chosen identifier for this datasource - independent of type, since multiple datasources can share one provider type.
          example: valorant
        type:
          type: string
          description: A provider type from GET /datasources's `available_types` list.
          example: http_sql
        config:
          type: object
          description: Provider-specific config; see GET /datasources for each provider type's config schema.
          additionalProperties: true
...